made UDependency inner class of CaptionToken
