+++ thumbnailAnnotator.parent/thumbnailAnnotator.core/src/main/java/nlp/floschne/thumbnailAnnotator/core/domain/CrawlerResult.java
@@ -14,4 +14,5 @@ import java.util.List;
 public class CrawlerResult extends DomainObject {
     private CaptionToken captionToken;
     private List<Thumbnail> thumbnails;
+
 }
